greenhouse_monitor/requirements.txt
pyserial~=3.5
google.cloud.logging
PyYAML~=6.0
grpcio
google-cloud-monitoring
pip==22.3.1
setuptools==65.6.3
